// Frontend/src/app/core/models/lancamento.model.ts
export class LancamentoModel {

  id: any;
  numeroProcesso: string;
  comarca: string;
  vara: string;
  dataEnvioEscritorio: Date;
  escritorio: string;
  tipoLancamento: string;
  categoriaPagamento: string;
  statusPagamento: string;
  dataLancamento: string;
  numeroPedidoSAP: string;
  dataRecebimentoFiscal: string;
  dataPagamentoPedido: string;
  valorLiquido: number;
  textoSAP: string;
  comentario: string;
  numeroGuia: string;
  autor: string;
  numeroContaJudicial: string;
  numeroParcelaJudicial: string;
  autenticacaoEletronica: string;
  statusParcelaBancoDoBrasil: string;
  dataEfetivacaoParcelaBancoDoBrasil: string;
  codigoProcesso: number;
  codigoLancamento: number;
  lancamentoEstornado: boolean;
  isEditavel: boolean;
  isDisabled: boolean;


}